+++ src/main/user-impl/java/com/mysql/cj/jdbc/ha/plugins/failover/FailoverConnectionPlugin.java
@@ -514,6 +514,13 @@ public class FailoverConnectionPlugin implements IConnectionPlugin {
     metricsContainer.registerReaderFailoverProcedureTime(currentTimeMs - this.failoverStartTimeMs);
     this.failoverStartTimeMs = 0;
 
+    if (result != null) {
+      final SQLException exception = result.getException();
+      if (exception != null) {
+        throw exception;
+      }
+    }
+
     if (result == null || !result.isConnected()) {
       // "Unable to establish SQL connection to reader node"
       processFailoverFailure(Messages.getString("ClusterAwareConnectionProxy.4"));
@@ -544,6 +551,13 @@ public class FailoverConnectionPlugin implements IConnectionPlugin {
     long currentTimeMs = System.currentTimeMillis();
     metricsContainer.registerWriterFailoverProcedureTime(currentTimeMs - this.failoverStartTimeMs);
     this.failoverStartTimeMs = 0;
+
+    if (failoverResult != null) {
+      final SQLException exception = failoverResult.getException();
+      if (exception != null) {
+        throw exception;
+      }
+    }
 
     if (failoverResult == null || !failoverResult.isConnected()) {
       // "Unable to establish SQL connection to writer node"
@@ -621,24 +635,22 @@ public class FailoverConnectionPlugin implements IConnectionPlugin {
       return false;
     }
 
-    String sqlState = null;
     if (t instanceof CommunicationsException || t instanceof CJCommunicationsException) {
       return true;
-    } else if (t instanceof SQLException) {
-      sqlState = ((SQLException) t).getSQLState();
-    } else if (t instanceof CJException) {
+    }
+
+    if (t instanceof SQLException) {
+      return ConnectionUtils.isNetworkException((SQLException) t);
+    }
+
+    if (t instanceof CJException) {
       if (t.getCause() instanceof EOFException) { // Can not read response from server
         return true;
       }
       if (t.getCause() instanceof SSLException) { // Incomplete packets from server may cause SSL communication issues
         return true;
       }
-      sqlState = ((CJException) t).getSQLState();
-    }
-
-    if (sqlState != null) {
-      // connection error
-      return sqlState.startsWith("08");
+      return ConnectionUtils.isNetworkException(((CJException) t).getSQLState());
     }
 
     return false;